REF: move QDataclassBridge to own submodule outside of atef.bin
